move join page /join/[slug]
